Move eventsService to src/lib/supabase directory
